--- tests/test_load_loss_functions.py
@@ -1,5 +1,0 @@
-from pytorch_optimizer import get_supported_loss_functions
-
-
-def test_get_supported_loss_functions():
-    assert len(get_supported_loss_functions()) == 10
